pyproject.toml
[tool.poetry]
name = "imagination"
version = "0.1.0"
description = ""
authors = ["eulerianKnight <himanshu3.1415926@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "^1.0.1"
langgraph = "^0.2.12"
langchain = "^0.2.14"
langchain-openai = "^0.1.22"
langchain-community = "^0.2.12"
tavily-python = "^0.4.0"
redis = "^5.0.8"
fastapi = "^0.112.1"
uvicorn = "^0.30.6"
chromadb = "^0.5.5"
langchain-chroma = ">=0.1.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
